+++ backend@nest/src/core/php/constants.ts
@@ -1,0 +1,20 @@
+/**
+ * express server
+ */
+export const PORT = 4000;
+export const HOST_NAME = 'localhost';
+
+/**
+ * for view engine
+ */
+export const TEMPLATE_FILES_PATH = '../../../../public/server/';
+
+
+/**
+ * i18n
+ */
+export const LANG = {
+    en: {
+        serverRun: '> PHP server listening at http://%s:%s'
+    }
+}