+++ backend@nest/src/core/php/php.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { PhpController } from './php.controller';
+import { PhpService } from './php.service';
+
+@Module({
+  imports: [],
+  controllers: [PhpController],
+  providers: [PhpService],
+})
+export class PhpModule {}